Add initial working code
Functional, and seems to give the same result as postrunner does.
However, I can't get some of the additional checks that postrunner makes
to succeed here, so they're currently left out. (For example, that
checks if the three bytes between the 2307 are zero, but I can't get
that check to pass, so maybe I'm doing something wrong...)